% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_ceres_results.R
\name{collate_ceres_results}
\alias{collate_ceres_results}
\title{Collate CERES results}
\usage{
collate_ceres_results(ceres_results, sg_data, cn_data)
}
\arguments{
\item{ceres_results}{List returned by \code{fit_ceres}}

\item{sg_data}{Matrix with sgRNA sequences as row names, samples (i.e. replicates of cell line screens) as column names}

\item{cn_data}{Matrix with genomic cut sites as row names, cell lines as column names}
}
\value{
A list of objects returned by the CERES fit
}
\description{
Collate CERES results
}
\examples{

}
