% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_seg_intersect.R
\name{intersect_guide_with_copy_number}
\alias{intersect_guide_with_copy_number}
\title{Get a matrix of CN for CRISPR target sites}
\usage{
intersect_guide_with_copy_number(guide.gr, seg.gr, CN.column = "CN",
  guide.column = "Guide", do_parallel = F)
}
\arguments{
\item{guide.gr}{GRanges object of sgRNA targets}

\item{seg.gr}{List of GRanges objects of CN segments}

\item{CN.column}{column to find CN data in seg.gr}

\item{guide.column}{guide.gr must contain a column with guide sequence}

\item{do_parallel}{run on multiple cores}
}
\value{
a matrix of CN data, each row is a guide, each column is a sample
}
\description{
Get a matrix of CN for CRISPR target sites
}
