% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utils.R
\name{mat.to.df}
\alias{mat.to.df}
\title{convert matrix to data frame}
\usage{
mat.to.df(mat, row.name = "Row", col.name = "Col", dat.name = "Dat")
}
\arguments{
\item{mat}{matrix}

\item{row.name}{column label of resulting data frame from \code{mat} rows}

\item{col.name}{column label of resulting data frame from \code{mat} cols}

\item{dat.name}{column label of resulting data frame from \code{mat} data}
}
\value{
a data frame
}
\description{
Converts a matrix to data frame
}
