% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utils.R
\name{remove.cols.all.nas}
\alias{remove.cols.all.nas}
\title{remove columns from a matrix that include any NAs}
\usage{
remove.cols.all.nas(x, indices = FALSE)
}
\arguments{
\item{x}{a matrix}

\item{indices}{logical, return indices to remove instead}
}
\value{
the matrix with columns removed, unless indices is TRUE
}
\description{
remove columns from a matrix that include any NAs
}
