% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ceres.R
\name{run_ceres}
\alias{run_ceres}
\title{CERES main routine}
\usage{
run_ceres(sg_data, cn_data, guide_locus, locus_gene, replicate_map, params)
}
\arguments{
\item{sg_data}{matrix with sgRNA sequences as row names, samples (i.e. replicates of cell line screens) as column names.}

\item{cn_data}{matrix with genomic cut sites as row names, cell lines as column names.}

\item{guide_locus}{data.frame with column names `Guide` and `Locus`. Values must match those corresponding to row names of sg_data and cn_data.}

\item{locus_gene}{data.frame with column names `Locus` and `Gene`.}

\item{replicate_map}{data.frame with column names `Replicate` and `CellLine`}

\item{params}{list of run parameters}
}
\value{
A list of numeric vectors and matrices containing the results of the CERES fit.
}
\description{
CERES main routine
}
