% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_to_essentials.R
\name{scale_to_essentials}
\alias{scale_to_essentials}
\title{Normalize CERES gene effects according to essential and nonessential genes}
\usage{
scale_to_essentials(ge_fit)
}
\arguments{
\item{ge_fit}{Matrix of gene effects by cell line. Expects row names are gene symbols and column names are cell line IDs.}
}
\value{
Matrix of scaled gene effects where the median of essential / nonessential genes are -1 and 0, respectively, for all cell lines.
}
\description{
Normalize CERES gene effects according to essential and nonessential genes
}
